requirements.txt
Flask==0.10.1
Flask-SocketIO==2.0
pyserial==3.0.1
